+++ src/Preferences/PreferencesDialog.gd
@@ -1,28 +1,28 @@
 extends AcceptDialog
 
-# Preferences table: [Prop name in Global, relative node path, value type]
+# Preferences table: [Prop name in Global, relative node path, value type, default value]
 var preferences = [
-	["open_last_project", "Startup/StartupContainer/OpenLastProject", "pressed"],
-	["smooth_zoom", "Canvas/SmoothZoom", "pressed"],
-	["pressure_sensitivity_mode", "Startup/PressureSentivity/PressureSensitivityOptionButton", "selected"],
-	["show_left_tool_icon", "Indicators/IndicatorsContainer/LeftToolIconCheckbox", "pressed"],
-	["show_right_tool_icon", "Indicators/IndicatorsContainer/RightToolIconCheckbox", "pressed"],
-	["left_square_indicator_visible", "Indicators/IndicatorsContainer/LeftIndicatorCheckbox", "pressed"],
-	["right_square_indicator_visible", "Indicators/IndicatorsContainer/RightIndicatorCheckbox", "pressed"],
-	["autosave_interval", "Backup/AutosaveContainer/AutosaveInterval", "value"],
-	["enable_autosave", "Backup/AutosaveContainer/EnableAutosave", "pressed"],
+	["open_last_project", "Startup/StartupContainer/OpenLastProject", "pressed", Global.open_last_project],
+	["smooth_zoom", "Canvas/ZoomOptions/SmoothZoom", "pressed", Global.smooth_zoom],
+	["pressure_sensitivity_mode", "Startup/PressureSentivity/PressureSensitivityOptionButton", "selected", Global.pressure_sensitivity_mode],
+	["show_left_tool_icon", "Indicators/IndicatorsContainer/LeftToolIconCheckbox", "pressed", Global.show_left_tool_icon],
+	["show_right_tool_icon", "Indicators/IndicatorsContainer/RightToolIconCheckbox", "pressed", Global.show_right_tool_icon],
+	["left_square_indicator_visible", "Indicators/IndicatorsContainer/LeftIndicatorCheckbox", "pressed", Global.left_square_indicator_visible],
+	["right_square_indicator_visible", "Indicators/IndicatorsContainer/RightIndicatorCheckbox", "pressed", Global.right_square_indicator_visible],
+	["autosave_interval", "Backup/AutosaveContainer/AutosaveInterval", "value", Global.autosave_interval],
+	["enable_autosave", "Backup/AutosaveContainer/EnableAutosave", "pressed", Global.enable_autosave],
 
-	["default_image_width", "Image/ImageOptions/ImageDefaultWidth", "value"],
-	["default_image_height", "Image/ImageOptions/ImageDefaultHeight", "value"],
-	["default_fill_color", "Image/ImageOptions/DefaultFillColor", "color"],
+	["default_image_width", "Image/ImageOptions/ImageDefaultWidth", "value", Global.default_image_width],
+	["default_image_height", "Image/ImageOptions/ImageDefaultHeight", "value", Global.default_image_height],
+	["default_fill_color", "Image/ImageOptions/DefaultFillColor", "color", Global.default_fill_color],
 
-	["grid_width", "Canvas/GridOptions/GridWidthValue", "value"],
-	["grid_height", "Canvas/GridOptions/GridHeightValue", "value"],
-	["grid_color", "Canvas/GridOptions/GridColor", "color"],
-	["guide_color", "Canvas/GuideOptions/GuideColor", "color"],
-	["checker_size", "Canvas/CheckerOptions/CheckerSizeValue", "value"],
-	["checker_color_1", "Canvas/CheckerOptions/CheckerColor1", "color"],
-	["checker_color_2", "Canvas/CheckerOptions/CheckerColor2", "color"],
+	["grid_width", "Canvas/GridOptions/GridWidthValue", "value", Global.grid_width],
+	["grid_height", "Canvas/GridOptions/GridHeightValue", "value", Global.grid_height],
+	["grid_color", "Canvas/GridOptions/GridColor", "color", Global.grid_color],
+	["guide_color", "Canvas/GuideOptions/GuideColor", "color", Global.guide_color],
+	["checker_size", "Canvas/CheckerOptions/CheckerSizeValue", "value", Global.checker_size],
+	["checker_color_1", "Canvas/CheckerOptions/CheckerColor1", "color", Global.checker_color_1],
+	["checker_color_2", "Canvas/CheckerOptions/CheckerColor2", "color", Global.checker_color_2],
 ]
 
 var selected_item := 0
@@ -30,6 +30,7 @@ var selected_item := 0
 onready var list : ItemList = $HSplitContainer/List
 onready var right_side : VBoxContainer = $HSplitContainer/ScrollContainer/VBoxContainer
 onready var autosave_interval : SpinBox = $HSplitContainer/ScrollContainer/VBoxContainer/Backup/AutosaveContainer/AutosaveInterval
+onready var restore_default_button_scene = preload("res://src/Preferences/RestoreDefaultButton.tscn")
 
 
 func _ready() -> void:
@@ -43,46 +44,64 @@ func _ready() -> void:
 
 	for pref in preferences:
 		var node = right_side.get_node(pref[1])
+		var node_position = node.get_index()
+
+		var restore_default_button : BaseButton = restore_default_button_scene.instance()
+		restore_default_button.setting_name = pref[0]
+		restore_default_button.value_type = pref[2]
+		restore_default_button.default_value = pref[3]
+		restore_default_button.node = node
+		node.get_parent().add_child(restore_default_button)
+		node.get_parent().move_child(restore_default_button, node_position)
 
 		match pref[2]:
 			"pressed":
-				node.connect("toggled", self, "_on_Preference_toggled", [pref[0]])
+				node.connect("toggled", self, "_on_Preference_toggled", [pref[0], pref[3], restore_default_button])
 			"value":
-				node.connect("value_changed", self, "_on_Preference_value_changed", [pref[0]])
+				node.connect("value_changed", self, "_on_Preference_value_changed", [pref[0], pref[3], restore_default_button])
 			"color":
 				node.get_picker().presets_visible = false
-				node.connect("color_changed", self, "_on_Preference_color_changed", [pref[0]])
+				node.connect("color_changed", self, "_on_Preference_color_changed", [pref[0], pref[3], restore_default_button])
 			"selected":
-				node.connect("item_selected", self, "_on_Preference_item_selected", [pref[0]])
+				node.connect("item_selected", self, "_on_Preference_item_selected", [pref[0], pref[3], restore_default_button])
 
 		if Global.config_cache.has_section_key("preferences", pref[0]):
 			var value = Global.config_cache.get_value("preferences", pref[0])
 			Global.set(pref[0], value)
 			node.set(pref[2], value)
 
+			# This is needed because color_changed doesn't fire if the color changes in code
+			if pref[2] == "color":
+				preference_update(pref[0])
+				disable_restore_default_button(restore_default_button, Global.get(pref[0]).is_equal_approx(pref[3]))
 
-func _on_Preference_toggled(button_pressed : bool, prop : String) -> void:
+
+func _on_Preference_toggled(button_pressed : bool, prop : String, default_value, restore_default_button : BaseButton) -> void:
 	Global.set(prop, button_pressed)
 	Global.config_cache.set_value("preferences", prop, button_pressed)
 	preference_update(prop)
+	disable_restore_default_button(restore_default_button, Global.get(prop) == default_value)
 
 
-func _on_Preference_value_changed(value : float, prop : String) -> void:
+func _on_Preference_value_changed(value : float, prop : String, default_value, restore_default_button : BaseButton) -> void:
 	Global.set(prop, value)
 	Global.config_cache.set_value("preferences", prop, value)
 	preference_update(prop)
+	disable_restore_default_button(restore_default_button, Global.get(prop) == default_value)
 
 
-func _on_Preference_color_changed(color : Color, prop : String) -> void:
+func _on_Preference_color_changed(color : Color, prop : String, default_value, restore_default_button : BaseButton) -> void:
 	Global.set(prop, color)
 	Global.config_cache.set_value("preferences", prop, color)
 	preference_update(prop)
+	disable_restore_default_button(restore_default_button, Global.get(prop).is_equal_approx(default_value))
 
 
-func _on_Preference_item_selected(id : int, prop : String) -> void:
+func _on_Preference_item_selected(id : int, prop : String, default_value, restore_default_button : BaseButton) -> void:
 	Global.set(prop, id)
 	Global.config_cache.set_value("preferences", prop, id)
 	preference_update(prop)
+	disable_restore_default_button(restore_default_button, Global.get(prop) == default_value)
 
 
 func preference_update(prop : String) -> void:
@@ -106,6 +125,16 @@ func preference_update(prop : String) -> void:
 				guide.default_color = Global.guide_color
 
 	Global.config_cache.save("user://cache.ini")
+
+
+func disable_restore_default_button(button : BaseButton, disable : bool) -> void:
+	button.disabled = disable
+	if disable:
+		button.mouse_default_cursor_shape = Control.CURSOR_ARROW
+		button.hint_tooltip = ""
+	else:
+		button.mouse_default_cursor_shape = Control.CURSOR_POINTING_HAND
+		button.hint_tooltip = "Restore default value"
 
 
 func _on_PreferencesDialog_about_to_show(changed_language := false) -> void: